backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RAG-powered search chatbot with standard and deep research modes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.28.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["oai", "tvly", "workflow"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line length handled by formatter
    "B008",   # do not perform function calls in argument defaults (FastAPI uses this)
    "E402",   # module level import not at top (needed for env var setup before imports)
    "SIM112", # env var capitalization (dotenv files may use lowercase)
]

[tool.ruff.lint.isort]
known-first-party = ["oai", "tvly", "workflow"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
